pyproject.toml
[project]
name = "property"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-cors>=4.0.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "ipykernel>=7.1.0",
    "ipynb>=0.5.1",
    "joblib>=1.4.2",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.45",
    "toml>=0.10.2",
]

[project.scripts]
property = "property.cli:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
